+++ test/unit/support/startup/appTemplateVars.test.js
@@ -27,7 +27,13 @@ test['app template vars'] = {
     yield this.shutdownApp();
   },
   'sets template vars': function*() {
-    this.App.routes = 'blah';
+    this.App.routes = {
+      url: {
+        bind: function() {
+          return 'blah';
+        }
+      }
+    };
     this.App.staticUrl = 'blah2';
 
     yield this.setup(this.App);