+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog for [waigo](https://github.com/waigo/waigo)
 
+## 24 Aug 2016 (2.1.9)
+* Fix `routeUrl` template var reference
+
 ## 22 Aug 2016 (2.1.8)
 * Ensure CLI commands run in app folder.
 * For expected CLI errors don't actually throw an error.